Add .pick and .assign methods to Runnable
